фикс вызова функции отрисовки экрана приветствия в хедере. модуль game-screen теперь экспортирует функцию отрисовки игровых экранов
